# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CTM Application api Documentation",
        "description": "L5 Swagger OpenApi description of CTM api",
        "contact": {
            "email": "tarikulislamnahid15@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/admin/employee/department/all/filtered": {
            "post": {
                "tags": [
                    "EMPLOYEE"
                ],
                "summary": "get paginated employee department from database",
                "description": "get paginated employee department from database",
                "operationId": "getAllEmpDepPaginated",
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "searchText": {
                                        "description": "search text for searching by employee department name",
                                        "type": "text"
                                    },
                                    "perPage": {
                                        "description": "number of employee department per page",
                                        "type": "text"
                                    },
                                    "page": {
                                        "description": "page number",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/admin/employee/department/insert": {
            "post": {
                "tags": [
                    "EMPLOYEE"
                ],
                "summary": "insert a Employee Department",
                "description": "insert a Employee Department",
                "operationId": "InsertEmpDep",
                "requestBody": {
                    "description": "enter inputs",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name of the Employee Department",
                                        "type": "text"
                                    },
                                    "description": {
                                        "description": "description of the Employee Department",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Successful Insert operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/admin/employee/department/update": {
            "post": {
                "tags": [
                    "EMPLOYEE"
                ],
                "summary": "update a employee department",
                "description": "update a employee department",
                "operationId": "UpdateEmpDep",
                "requestBody": {
                    "description": "enter inputs",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "id of the employee department",
                                        "type": "text"
                                    },
                                    "name": {
                                        "description": "name of the employee department",
                                        "type": "text"
                                    },
                                    "description": {
                                        "description": "description of the employee department",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Successful Insert operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/admin/employee/department/delete": {
            "post": {
                "tags": [
                    "EMPLOYEE"
                ],
                "summary": "delete a employee department",
                "description": "delete a employee department",
                "operationId": "DeleteEmpDep",
                "requestBody": {
                    "description": "enter inputs",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "id of the employee department",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Successful Insert operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}